Expansion handler added
Ideally to serve as a way to handle english and portuguese differently
